refactor: extract analyst-annotation routes into routes/findings.ts
Pure structural move out of createApp; no handler logic changed. Moves the
false-positive (list/mark/batch/remove/suggest), scope, correlation-profile,
comments, tags, pinned-findings, and notebook routes. No RouteContext members
were graduated: the domain reuses the already-graduated resynthesizeInBackground
and the stable ctx surface (store/options/hasAiProvider). FalsePositiveStore and
ScopeStore are stateless disk-backed, so they are rebuilt in-module (identical
construction over ctx.store, the threatIntel/anonymization precedent); createApp
keeps its own FalsePositiveStore for the whitelist/NSRL auto-sweeps that stay.
confidence-control is left in createApp for the aiSynthesis domain.
